+++ Source/RxSwift/mapTo.swift
@@ -17,7 +17,18 @@ extension ObservableType {
 	- parameter value: A constant that each element of the input sequence is being replaced with
 	- returns: An observable sequence containing the values `value` provided as a parameter
 	*/
-	public func mapTo<R>(_ value: R) -> Observable<R> {
+    @available(*, deprecated, renamed: "map(to:)")
+    public func mapTo<R>(_ value: R) -> Observable<R> {
+        return map(to: value)
+    }
+
+    /**
+     Returns an observable sequence containing as many elements as its input but all of them are the constant provided as a parameter
+
+     - parameter value: A constant that each element of the input sequence is being replaced with
+     - returns: An observable sequence containing the values `value` provided as a parameter
+     */
+	public func map<R>(to value: R) -> Observable<R> {
 		return map { _ in value }
 	}
 }